[SharedSE] Move CSSE address tables into SharedSE, named by target binary
Per the architectural guidance: address tables should be named by the
target executable they describe, not by the project that owns them. This
lets future consumers (MWSE, MGE-XE etc.) just include the file matching
their target binary without project coupling.

- New: SharedSE/NIConfig.TESConstructionSet.h (verbatim move of
  CSSE/NIConfig.h content; dropped redundant `#include NIDefines.h`
  since SharedSE/NIDefines.h includes us, not the other way around).
- New: SharedSE/NIRTTIDefines.TESConstructionSet.h (verbatim move of
  CSSE/NIRTTIDefines.h content).
- CSSE/NIConfig.h is now a 1-line shim including the per-target file.
- CSSE/NIRTTIDefines.h is now a 1-line shim too.
- SharedSE.vcxitems registers the two new headers.

Verified: CSSE Release|x86 builds clean. CSSE.dll size unchanged
(2861056 bytes); content hash differs only in PE timestamp.

The same per-target pattern will be applied to Morrowind addresses in
the next commit (NIConfig.Morrowind.h + NIRTTIDefines.Morrowind.h),
unblocking the MWSE side of the SharedSE consumption.
