+++ CSSE/NIRTTIDefines.h
@@ -1,116 +1,4 @@
 #pragma once
 
-namespace NI::RTTIStaticPtr {
-	enum RTTIStaticPtr : uintptr_t {
-		AvoidNode = 0x6D69C8,
-		BSMirroredNode = 0x6D06E8,
-		NiAccumulator = 0x6D8104,
-		NiAlphaAccumulator = 0x6D72F0,
-		NiAlphaController = 0x6D836C,
-		NiAlphaProperty = 0x6D7324,
-		NiAmbientLight = 0x6D7334,
-		NiAutoNormalParticles = 0x6D740C,
-		NiAutoNormalParticlesData = 0x6D73EC,
-		NiAVObject = 0x6D7A2C,
-		NiBillboardNode = 0x6D7448,
-		NiBltSource = 0x6D74B0,
-		NiBSAnimationManager = 0x6D7E14,
-		NiBSAnimationNode = 0x6D7E38,
-		NiBSPArrayController = 0x6D8C40,
-		NiBSParticleNode = 0x6D800C,
-		NiBSPNode = 0x6D74E4,
-		NiCamera = 0x6D72A0,
-		NiClusterAccumulator = 0x6D72C8,
-		NiCollisionSwitch = 0x6D74F8,
-		NiColorData = 0x6D83CC,
-		NiDirectionalLight = 0x6D7514,
-		NiDitherProperty = 0x6D753C,
-		NiDX8Renderer = 0x6D6B68,
-		NiDynamicEffect = 0x6D7C80,
-		NiExtraData = 0x6D71E4,
-		NiFlipController = 0x6D8404,
-		NiFloatController = 0x6D8CF0,
-		NiFloatData = 0x6D8350,
-		NiFltAnimationNode = 0x6D7590,
-		NiFogProperty = 0x6D7598,
-		NiGeometry = 0x6D7C1C,
-		NiGeometryData = 0x6D7E78,
-		NiGeomMorpherController = 0x6D844C,
-		NiGravity = 0x6D847C,
-		NiKeyframeController = 0x6D82CC,
-		NiKeyframeData = 0x6D8294,
-		NiKeyframeManager = 0x6D8324,
-		NiLight = 0x6D7DE8,
-		NiLightColorController = 0x6D84C4,
-		NiLines = 0x6D75D0,
-		NiLinesData = 0x6D75B4,
-		NiLODNode = 0x6D7618,
-		NiLookAtController = 0x6D8510,
-		NiMaterialColorController = 0x6D8544,
-		NiMaterialProperty = 0x6D7630,
-		NiMorphData = 0x6D8414,
-		NiMorpherController = 0x6D8CB4,
-		NiNode = 0x6D7260,
-		NiObject = 0x6D7A3C,
-		NiObjectNET = 0x6D71F8,
-		NiPalette = 0x6D746C,
-		NiParticleBomb = 0x6D8578,
-		NiParticleCollider = 0x6D8D54,
-		NiParticleColorModifier = 0x6D85A8,
-		NiParticleGrowFade = 0x6D85D0,
-		NiParticleModifier = 0x6D8D20,
-		NiParticleRotation = 0x6D860C,
-		NiParticles = 0x6D73AC,
-		NiParticlesData = 0x6D7394,
-		NiParticleSystemController = 0x6D8230,
-		NiPathController = 0x6D8660,
-		NiPixelData = 0x6D7494,
-		NiPlanarCollider = 0x6D8698,
-		NiPointLight = 0x6D7650,
-		NiPosData = 0x6D84A0,
-		NiProperty = 0x6D7C40,
-		NiRenderedCubeMap = 0x6D806C,
-		NiRenderedTexture = 0x6D7664,
-		NiRenderer = 0x6D7EDC,
-		NiRendererSpecificProperty = 0x6D7698,
-		NiRollController = 0x6D86AC,
-		NiRotatingParticles = 0x6D76CC,
-		NiRotatingParticlesData = 0x6D76B0,
-		NiScreenPolygon = 0x6D7288,
-		NiSequenceStreamHelper = 0x6D86C4,
-		NiShadeProperty = 0x6D770C,
-		NiSkinData = 0x6D7364,
-		NiSkinInstance = 0x6D7378,
-		NiSkinPartition = 0x6D722C,
-		NiSortAdjustNode = 0x6D7730,
-		NiSourceTexture = 0x6D7768,
-		NiSpecularProperty = 0x6D7784,
-		NiSphericalCollider = 0x6D86EC,
-		NiSpotLight = 0x6D77A8,
-		NiStencilProperty = 0x6D77C0,
-		NiStringExtraData = 0x6D77D4,
-		NiSwitchNode = 0x6D7554,
-		NiTextKeyExtraData = 0x6D8304,
-		NiTexture = 0x6D7A78,
-		NiTextureEffect = 0x6D782C,
-		NiTexturingProperty = 0x6D77E8,
-		NiTimeController = 0x6D7DC4,
-		NiTriBasedGeom = 0x6D7CB4,
-		NiTriBasedGeomData = 0x6D812C,
-		NiTriShape = 0x6D78F8,
-		NiTriShapeData = 0x6D7838,
-		NiTriShapeDynamicData = 0x6D7910,
-		NiTriStrips = 0x6D7950,
-		NiTriStripsData = 0x6D7930,
-		NiUVController = 0x6D8770,
-		NiUVData = 0x6D8748,
-		NiVertexColorProperty = 0x6D7978,
-		NiVertWeightsExtraData = 0x6D7990,
-		NiVisController = 0x6D87A4,
-		NiVisData = 0x6D878C,
-		NiWireframeProperty = 0x6D79AC,
-		NiZBufferProperty = 0x6D79D8,
-		RootCollisionNode = 0x6D69E0,
-		TES3ObjectExtraData = 0x6D5728,
-	};
-}
+// CSSE targets TESConstructionSet.exe.
+#include "NIRTTIDefines.TESConstructionSet.h"
