+++ SharedSE/NIConfig.TESConstructionSet.h
@@ -1,0 +1,143 @@
+#pragma once
+
+// Engine function addresses for TESConstructionSet.exe (the Construction Set editor binary).
+// Consumers select the appropriate target via a thin per-project NIConfig.h shim.
+
+// NI::Matrix33
+#define SE_NI_MATRIX33_FNADDR_ADDMATRIX 0x5E1F20
+#define SE_NI_MATRIX33_FNADDR_FROMEULERXYZ 0x5E3760
+#define SE_NI_MATRIX33_FNADDR_INVERSE 0x5E23A0
+#define SE_NI_MATRIX33_FNADDR_INVERSERAW 0x5E22B0
+#define SE_NI_MATRIX33_FNADDR_MULTIPLYMATRIX 0x5E2060
+#define SE_NI_MATRIX33_FNADDR_MULTIPLYSCALAR 0x5E2170
+#define SE_NI_MATRIX33_FNADDR_MULTIPLYVECTOR 0x5E21F0
+#define SE_NI_MATRIX33_FNADDR_REORTHOGONALIZE 0x5E26D0
+#define SE_NI_MATRIX33_FNADDR_SUBTRACTMATRIX 0x5E1FC0
+#define SE_NI_MATRIX33_FNADDR_TESTEQUAL 0x0
+#define SE_NI_MATRIX33_FNADDR_TOEULERXYZ 0x5E3110
+#define SE_NI_MATRIX33_FNADDR_TOIDENTITY 0x5E1C60
+#define SE_NI_MATRIX33_FNADDR_TOROTATIONX 0x5E1CC0
+#define SE_NI_MATRIX33_FNADDR_TOROTATIONXYZ 0x5E1D80
+#define SE_NI_MATRIX33_FNADDR_TOROTATIONY 0x5E1D00
+#define SE_NI_MATRIX33_FNADDR_TOROTATIONZ 0x5E1D40
+#define SE_NI_MATRIX33_FNADDR_TRANSPOSE 0x5E23E0
+
+// NI::Quaternion
+#define SE_NI_QUATERNION_FNADDR_FROMANGLEAXIS 0x0
+#define SE_NI_QUATERNION_FNADDR_FROMROTATION 0x0
+#define SE_NI_QUATERNION_FNADDR_INVERT 0x0
+#define SE_NI_QUATERNION_FNADDR_MULTIPLYQUATERNION 0x0
+#define SE_NI_QUATERNION_FNADDR_SLERP 0x0
+#define SE_NI_QUATERNION_FNADDR_TOANGLEAXIS 0x0
+
+// NI::Object
+#define SE_NI_OBJECT_FNADDR_CREATECLONE 0x5DCC70
+#define SE_NI_OBJECT_FNADDR_CTOR 0x5B27B0
+#define SE_NI_OBJECT_FNADDR_DTOR 0x5B2860
+
+// NI::ObjectNET
+#define SE_NI_OBJECTNET_FNADDR_ADDEXTRADATA 0x5B2AA0
+#define SE_NI_OBJECTNET_FNADDR_PREPENDCONTROLLER 0x0
+#define SE_NI_OBJECTNET_FNADDR_REMOVEALLCONTROLLERS 0x0
+#define SE_NI_OBJECTNET_FNADDR_REMOVEALLEXTRADATA 0x0
+#define SE_NI_OBJECTNET_FNADDR_REMOVECONTROLLER 0x0
+#define SE_NI_OBJECTNET_FNADDR_REMOVEEXTRADATA 0x5B2B50
+#define SE_NI_OBJECTNET_FNADDR_SETFLAG 0x44A130
+#define SE_NI_OBJECTNET_FNADDR_SETNAME 0x5B2A80
+
+// NI::AVObject
+#define SE_NI_AVOBJECT_FNADDR_ATTACHPROPERTY 0x44CF00
+#define SE_NI_AVObject_FNADDR_DETACHPROPERTYBYTYPE 0x5DADD0
+#define SE_NI_AVOBJECT_FNADDR_SETLOCALROTATIONMATRIX 0x462C70
+#define SE_NI_AVOBJECT_FNADDR_UPDATE 0x5DAFB0
+#define SE_NI_AVOBJECT_FNADDR_UPDATEEFFECTS 0x5DB340
+#define SE_NI_AVOBJECT_FNADDR_UPDATEPROPERTIES 0x5DB090
+
+// NI::Node
+#define SE_NI_NODE_FNADDR_ATTACHEFFECT 0x0
+#define SE_NI_NODE_FNADDR_CTOR 0x5B6100
+#define SE_NI_NODE_FNADDR_DETACHALLEFFECTS 0x0
+#define SE_NI_NODE_FNADDR_DETACHEFFECT 0x0
+
+// NI::DynamicEffect
+#define SE_NI_DYNAMICEFFECT_FNADDR_ATTACHAFFECTEDNODE 0x0
+#define SE_NI_DYNAMICEFFECT_FNADDR_CTOR 0x0
+#define SE_NI_DYNAMICEFFECT_FNADDR_DETACHAFFECTEDNODE 0x0
+#define SE_NI_FNADDR_CLEARDYNAMICEFFECTNODES 0x402824
+
+// NI::Property
+#define SE_NI_PROPERTY_FNADDR_CTOR 0x44A170
+#define SE_NI_PROPERTY_FNADDR_DTOR 0x0
+#define SE_NI_PROPERTY_FNADDR_SETFLAGBITFIELD 0x0
+
+// NI::AlphaProperty
+#define SE_NI_ALPHAPROPERTY_VTBL 0x6732E8
+
+// NI::TexturingProperty::Map
+#define SE_NI_TEXTURINGPROPERTY_MAP_FNADDR_CTOR 0x0
+#define SE_NI_TEXTURINGPROPERTY_MAP_FNADDR_CTORWITHPARAMS 0x0
+
+// NI::TexturingProperty::BumpMap
+#define SE_NI_TEXTURINGPROPERTY_BUMPMAP_VTBL 0x0
+
+// NI::TexturingProperty
+#define SE_NI_TEXTURINGPROPERTY_FNADDR_CTOR 0x5D2F10
+#define SE_NI_TEXTURINGPROPERTY_FNADDR_DTOR 0x5D30D0
+
+// NI::VertexColorProperty
+#define SE_NI_VERTEXCOLORPROPERTY_VTBL 0x673170
+
+// NI::ZBufferProperty
+#define SE_NI_ZBUFFERPROPERTY_VTBL 0x6731F0
+
+// NI::SkinInstance
+#define SE_NI_SKININSTANCE_FNADDR_DEFORM 0x5F08A0
+
+// NI::TriShape
+#define SE_NI_TRISHAPE_FNADDR_CTOR 0x05D6D10
+
+// NI::TriShapeData
+#define SE_NI_TRISHAPEDATA_FNADDR_CREATE 0x5D6260
+
+// NI::Lines
+#define SE_NI_LINES_FNADDR_CTOR 0x5CB2A0
+
+// NI::GeometryData
+#define SE_NI_GEOMETRYDATA_FNADDR_LOADBINARY 0x5ED2C0
+
+// NI::StringExtraData
+#define SE_NI_STRINGEXTRADATA_FNADDR_CTOR 0x5D25F0
+#define SE_NI_STRINGEXTRADATA_FNADDR_DTOR 0x5D2D80
+
+// NI::Camera
+#define SE_NI_CAMERA_FNADDR_CLEAR 0x5BAB70
+#define SE_NI_CAMERA_FNADDR_CLICK 0x5BABA0
+#define SE_NI_CAMERA_FNADDR_CTOR 0x5BA5F0
+#define SE_NI_CAMERA_FNADDR_DTOR 0x5BA900
+#define SE_NI_CAMERA_FNADDR_WINDOWPOINTTORAY 0x5BC680
+#define SE_NI_CAMERA_FNADDR_WORLDPOINTTOSCREENPOINT 0x5BB870
+
+// NI::Pick
+#define SE_NI_PICK_FNADDR_CLEARRESULTS 0x5B34E0
+#define SE_NI_PICK_FNADDR_CTOR 0x5B3350
+#define SE_NI_PICK_FNADDR_DTOR 0x5B3400
+#define SE_NI_PICK_FNADDR_PICKOBJECTS 0x5B35B0
+
+// NI::Stream
+#define SE_NI_STREAM_FNADDR_CTOR 0x0
+#define SE_NI_STREAM_FNADDR_DTOR 0x0
+#define SE_NI_STREAM_FNADDR_GETCOPIEDOBJECT 0x0
+#define SE_NI_STREAM_FNADDR_GETLINKID 0x0
+#define SE_NI_STREAM_FNADDR_GETLINKOBJECT 0x0
+#define SE_NI_STREAM_FNADDR_INSERTOBJECT 0x0
+#define SE_NI_STREAM_FNADDR_LOAD 0x0
+#define SE_NI_STREAM_FNADDR_READSTRING 0x0
+#define SE_NI_STREAM_FNADDR_REGISTERLOADER 0x0
+#define SE_NI_STREAM_FNADDR_SAVE 0x0
+#define SE_NI_STREAM_FNADDR_WRITESTRING 0x0
+
+// NI::Stream::LoadingObject
+#define SE_NI_STREAM_LOADINGOBJECT_VTBL 0x0
+
+// NI::Bound
+#define SE_NI_BOUND_FNADDR_COMPUTEFROMDATA 0x5B13F0
